extract panel component
