+++ App/Components/ChannelPanels/index.js
@@ -1,37 +1,15 @@
 // @flow
 import React from 'react'
-import { Text, View, ScrollView, Image } from 'react-native'
+import { View, ScrollView } from 'react-native'
 import { chunk } from 'lodash'
 
 import type {TChannelStore} from '../../types/Channel'
-import Icon from 'react-native-vector-icons/FontAwesome'
+import {Panel} from './Panel'
 import styles from './style'
 
 type TChannelPanels = {
   channels: TChannelStore[],
 }
-
-type TChannelPanel = {
-  channel: TChannelStore,
-  isMargin: boolean
-}
-
-export const Panel = ({channel, isMargin}: TChannelPanel) => (
-  <View style={styles.panel}>
-    <View style={isMargin ? styles.panelContentEven : styles.panelContentOdd}>
-      <Image style={styles.thumbnail} source={{uri: `${channel.youtube.thumbnail}`}} />
-      <View style={styles.rank}>
-        <Text style={styles.rankText}>{channel.rank}</Text>
-      </View>
-      <View style={styles.panelInfo}>
-        <Text style={styles.name}>{channel.youtube.name}</Text>
-        <View style={styles.icon}>
-          <Icon name='rocket' size={20} color={channel.isLiked ? '#900' : '#ddd'} />
-        </View>
-      </View>
-    </View>
-  </View>
-)
 
 export const ChannelPanels = ({channels}: TChannelPanels) => (
   <View style={styles.channelPanels}>
