+++ App/Services/index.js
@@ -1,3 +1,5 @@
+// @flow
 export * from './firebaseRef'
 export * from './resources'
 export * from './batches'
+export * from './styles'
